requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.2.0
async-timeout==4.0.3
attrs==24.2.0
blas==1.0
blinker==1.6.2
brotli-bin==1.0.9
brotli-python==1.0.9
brotli==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2025.6.15
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.0.5
cryptography==43.0.0
easydict==1.13
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.0
fsspec==2024.6.1
gitdb==4.0.12
gitpython==3.1.44
google-auth-oauthlib==1.0.0
google-auth==2.40.3
grpcio==1.70.0
hf-xet==1.1.4
idna==3.7
importlib-metadata==8.5.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.4
markdown==3.7
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
oauthlib==3.3.0
opencv-python==4.11.0.86
packaging==24.1
pillow==10.4.0
protobuf==5.29.5
psutil==7.0.0
pydantic-core==2.27.2
pydantic==2.10.6
pyjwt==2.8.0
pyopenssl==24.2.1
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.9.0post0
pytorch==2.4.1
pyyaml==6.0.2
requests-oauthlib==2.0.0
requests==2.32.3
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.10.1
setproctitle==1.3.6
six==1.16.0
sympy==1.13.3
tensorboard==2.14.0
timm==1.0.15
torchaudio==2.4.1
tqdm==4.66.5
typing-extensions==4.13.2
typing_extensions==4.11.0
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2